+++ src/main/java/com/uber/cadence/internal/replay/DecisionContext.java
@@ -43,8 +43,7 @@ public interface DecisionContext extends ReplayAware {
 
   WorkflowExecution getWorkflowExecution();
 
-  // TODO: Add to Cadence
-  //    com.uber.cadence.WorkflowExecution getParentWorkflowExecution();
+  WorkflowExecution getParentWorkflowExecution();
 
   WorkflowType getWorkflowType();
 
@@ -87,11 +86,7 @@ public interface DecisionContext extends ReplayAware {
    */
   Map<String, Object> getPropagatedContexts();
 
-  /**
-   * Returns the set of configured context propagators
-   *
-   * @return
-   */
+  /** Returns the set of configured context propagators */
   List<ContextPropagator> getContextPropagators();
 
   /**