+++ src/main/java/com/uber/cadence/internal/sync/DeterministicRunnerImpl.java
@@ -240,7 +240,7 @@ class DeterministicRunnerImpl implements DeterministicRunner {
         toExecuteInWorkflowThread.clear();
         progress = false;
         Iterator<WorkflowThread> ci = threads.iterator();
-        nextWakeUpTime = 0;
+        nextWakeUpTime = Long.MAX_VALUE;
         while (ci.hasNext()) {
           WorkflowThread c = ci.next();
           progress = c.runUntilBlocked() || progress;
@@ -256,7 +256,7 @@ class DeterministicRunnerImpl implements DeterministicRunner {
             }
           } else {
             long t = c.getBlockedUntil();
-            if (t > nextWakeUpTime) {
+            if (t > currentTimeMillis() && t < nextWakeUpTime) {
               nextWakeUpTime = t;
             }
           }
@@ -269,7 +269,8 @@ class DeterministicRunnerImpl implements DeterministicRunner {
           threads.addLast(c);
         }
       } while (progress && !threads.isEmpty());
-      if (nextWakeUpTime < currentTimeMillis()) {
+
+      if (nextWakeUpTime < currentTimeMillis() || nextWakeUpTime == Long.MAX_VALUE) {
         nextWakeUpTime = 0;
       }
     } finally {
@@ -521,6 +522,11 @@ class DeterministicRunnerImpl implements DeterministicRunner {
 
     @Override
     public WorkflowExecution getWorkflowExecution() {
+      throw new UnsupportedOperationException("not implemented");
+    }
+
+    @Override
+    public WorkflowExecution getParentWorkflowExecution() {
       throw new UnsupportedOperationException("not implemented");
     }