+++ src/main/java/com/uber/cadence/internal/testservice/StateMachines.java
@@ -152,16 +152,22 @@ class StateMachines {
     int backoffStartIntervalInSeconds;
     String cronSchedule;
     byte[] lastCompletionResult;
+    String originalExecutionRunId;
+    Optional<String> continuedExecutionRunId;
 
     WorkflowData(
         Optional<RetryState> retryState,
         int backoffStartIntervalInSeconds,
         String cronSchedule,
-        byte[] lastCompletionResult) {
+        byte[] lastCompletionResult,
+        String originalExecutionRunId,
+        Optional<String> continuedExecutionRunId) {
       this.retryState = retryState;
       this.backoffStartIntervalInSeconds = backoffStartIntervalInSeconds;
       this.cronSchedule = cronSchedule;
       this.lastCompletionResult = lastCompletionResult;
+      this.originalExecutionRunId = originalExecutionRunId;
+      this.continuedExecutionRunId = continuedExecutionRunId;
     }
   }
 
@@ -539,10 +545,20 @@ class StateMachines {
     if (data.retryState.isPresent()) {
       a.setAttempt(data.retryState.get().getAttempt());
     }
+    a.setOriginalExecutionRunId(data.originalExecutionRunId);
+    if (data.continuedExecutionRunId.isPresent()) {
+      a.setContinuedExecutionRunId(data.continuedExecutionRunId.get());
+    }
     a.setLastCompletionResult(data.lastCompletionResult);
     a.setMemo(request.getMemo());
     a.setSearchAttributes((request.getSearchAttributes()));
     a.setHeader(request.getHeader());
+    Optional<TestWorkflowMutableState> parent = ctx.getWorkflowMutableState().getParent();
+    if (parent.isPresent()) {
+      ExecutionId parentExecutionId = parent.get().getExecutionId();
+      a.setParentWorkflowDomain(parentExecutionId.getDomain());
+      a.setParentWorkflowExecution(parentExecutionId.getExecution());
+    }
     HistoryEvent event =
         new HistoryEvent()
             .setEventType(EventType.WorkflowExecutionStarted)
@@ -717,6 +733,7 @@ class StateMachines {
 
     PollForActivityTaskResponse taskResponse =
         new PollForActivityTaskResponse()
+            .setWorkflowType(data.startWorkflowExecutionRequest.workflowType)
             .setActivityType(d.getActivityType())
             .setWorkflowExecution(ctx.getExecution())
             .setActivityId(d.getActivityId())